+++ CHANGES.rst
@@ -5,6 +5,10 @@ Changelog
 1.2.5 (unreleased)
 ------------------
 
+- WEB-4007 : Get ContactProperties out of imio.smartweb.core to also use it in imio.directory.core
+  and simplifying formated schedule displaying in REACT directory view
+  [boulch]
+
 - WEB-4029 : File and Image content types don't have WF so we set effective date equal to created date
   [boulch]